pyproject.toml
[project]
name = "apeiria_bot"
version = "0.1.0"
description = "apeiria_bot"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "click>=8.1.8",
    "nb-cli>=1.7.2",
    "packaging>=24.2",
    "nonebot2[aiohttp,fastapi]>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-localstore>=0.7.4",
    "nonebot-plugin-alconna>=0.61.0",
    "nonebot-plugin-orm[default]>=0.8.3",
    "playwright>=1.54.0",
    "pydantic>=2.0,<3.0",
    "PyJWT>=2.10.1",
    "tomlkit>=0.13.2",
]

[project.scripts]
apeiria = "apeiria.interfaces.cli.main:main"

[project.entry-points."nb_scripts"]
apeiria = "apeiria.interfaces.cli.main:main"

[tool.nonebot]
plugin_dirs = []
builtin_plugins = []

[tool.nonebot.adapters]

[tool.nonebot.plugins]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["apeiria*"]

[dependency-groups]
dev = [
    "pyright[nodejs]",
    "ruff",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
# For more rules, see https://docs.astral.sh/ruff/rules/.
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "I",     # isort
    "C90",   # mccabe
    "N",     # pep8-naming
    "PL",    # pylint
    "UP",    # pyupgrade
    "YTT",   # flake8-2020
    "ANN",   # flake8-annotations
    "ASYNC", # flake8-async
    "BLE",   # flake8-blind-except
    "FBT",   # flake8-boolean-trap
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "COM",   # flake8-commas
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "T10",   # flake8-debugger
    "ICN",   # flake8-import-conventions
    "PIE",   # flake8-pie
    "T20",   # flake8-print
    "PYI",   # flake8-pyi
    "Q",     # flake8-quotes
    "RSE",   # flake8-raise
    "RET",   # flake8-return
    "SIM",   # flake8-simplify
    "SLOT",  # flake8-slots
    "TID",   # flake8-tidy-imports
    "TC",    # flake8-type-checking
    "ARG",   # flake8-unused-arguments
    "PTH",   # flake8-use-pathlib
    # "ERA",   # eradicate
    "FAST",  # FastAPI
    "PERF",  # Perflint
    "PGH",   # pygrep-hooks
    "FURB",  # refurb
    "TRY",   # tryceratops
    "RUF",   # Ruff-specific rules
]
ignore = [
    "E402",    # module-import-not-at-top-of-file  # nonebot2 require() violates this
    "B008",    # function-call-in-default-argument  # nonebot2 Depends() without Annotated violates this
    "UP037",   # quoted-annotation
    # "RUF001",  # ambiguous-unicode-character-string
    # "RUF002",  # ambiguous-unicode-character-docstring
    # "RUF003",  # ambiguous-unicode-character-comment
    # "ANN201",  # missing-return-type-undocumented-public-function
    "ANN202",  # missing-return-type-private-function
    "ANN401",  # any-type
    "COM812",  # missing-trailing-comma
    "PLC0415", # import-outside-top-level
]
allowed-confusables = ["，", "。", "“", "”", "：", "；", "？", "！", "【", "】", "《", "》", "…", "—", "（", "）", "、"]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pyright]
# For Pylance/Pyright configurations, see https://microsoft.github.io/pyright/#/configuration.
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "standard"
extraPaths = ["."]
venvPath = "."
venv = ".venv"
